FloatConversion

+++ src/TCode.h
@@ -218,7 +218,7 @@ private:
     void runSetupCommand(TCode_Setup_Command &command);
     void runExternalCommand(TCode_External_Command &command);
 
-    void setSaveValues(TCode_ChannelID &id, unsigned int min, unsigned int max);
+    void setSaveValues(TCode_ChannelID &id, float min, float max);
 
 
     void printSavedAxisValues();
